requirements.txt
Flask>=3.0,<4.0
google-auth>=2.35.0
google-cloud-speech>=2.27.0
langchain-core>=0.3.0
langchain-classic>=0.3.0
langchain-community>=0.3.0
llama-cpp-python>=0.3.0
waitress>=3.0.0
